More pointer stuff

+++ src/include/duckdb/execution/ht_entry.hpp
@@ -40,22 +40,22 @@ public:
 	// Returns a pointer based on the stored value without checking cell occupancy.
 	// This can return a nullptr if the cell is not occupied.
 	inline data_ptr_t GetPointerOrNull() const {
-		return reinterpret_cast<data_ptr_t>(value & POINTER_MASK);
+		return reinterpret_cast<data_ptr_t>(static_cast<uintptr_t>(value & POINTER_MASK));
 	}
 
 	// Returns a pointer based on the stored value if the cell is occupied
 	inline data_ptr_t GetPointer() const {
 		D_ASSERT(IsOccupied());
-		return reinterpret_cast<data_ptr_t>(value & POINTER_MASK);
+		return reinterpret_cast<data_ptr_t>(static_cast<uintptr_t>(value & POINTER_MASK));
 	}
 
 	inline void SetPointer(const data_ptr_t &pointer) {
 		// Pointer shouldn't use upper bits
-		D_ASSERT((reinterpret_cast<uint64_t>(pointer) & SALT_MASK) == 0);
+		D_ASSERT((static_cast<uint64_t>(reinterpret_cast<uintptr_t>(pointer)) & SALT_MASK) == 0);
 		// Value should have all 1's in the pointer area
 		D_ASSERT((value & POINTER_MASK) == POINTER_MASK);
 		// Set upper bits to 1 in pointer so the salt stays intact
-		value &= reinterpret_cast<uint64_t>(pointer) | SALT_MASK;
+		value &= static_cast<uint64_t>(reinterpret_cast<uintptr_t>(pointer)) | SALT_MASK;
 	}
 
 	// Returns the salt, leaves upper salt bits intact, sets lower bits to all 1's
